Fix issue #3
This fixes the issue of overflow in the number of entries in each histogram (issue #3).

+++ include/histogram/Histogram1D.h
@@ -124,7 +124,7 @@ private:
   const Axis xaxis;
 
   //! The number of entries in the histogram.
-  int entries;
+  size_t entries;
 
   //! The bin contents, including the overflow bins.
   data_t *data;
