+++ include/histogram/Histogram3D.h
@@ -169,7 +169,7 @@ private:
     const Axis zaxis;
 
     //! The number of entries in the histogram.
-    int entries;
+    size_t entries;
 
 #ifndef USE_ROWS
     //! The bin contents, including the overflow bins.
